Move R8 sprite loading into Mods.D2k.
